+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 3.1.1
+
+### Fixed
+
+- Fix text in expired Gel notification's title and body - was previously the Injection title and body.
+- Fix text in all expired hormone notifications where it would indicate the hormone is not expired yet.
+- Fixed migration issue where you would get an extra pill time even if you were only using a single time.
+
+### Removed
+
+- Removed next suggested site text from expired hormone notifications, as it was faulty.
+
 ## 3.1.0
 
 ### Added